Repackage files into individual modules for easier use
