+++ AlignFive/interface/board.py
@@ -4,7 +4,7 @@ from typing import List
 
 import numpy as np
 
-from AlignFive.game_interface import GameWindow
+from AlignFive.interface.game_window import GameWindow
 from AlignFive.utils import Position, Move, PositionIndex
 
 if typing.TYPE_CHECKING:
